requirements.txt
Flask==2.0.1
pandas==1.3.3
fpdf2==2.5.6
python-dotenv==0.19.0
gunicorn==20.1.0
email-validator==1.1.3
pytest==7.2.0
pytest-cov==4.0.0